+++ src/oci/container_instances/models/linux_security_context.py
@@ -42,13 +42,18 @@ class LinuxSecurityContext(SecurityContext):
             The value to assign to the is_root_file_system_readonly property of this LinuxSecurityContext.
         :type is_root_file_system_readonly: bool
 
+        :param capabilities:
+            The value to assign to the capabilities property of this LinuxSecurityContext.
+        :type capabilities: oci.container_instances.models.ContainerCapabilities
+
         """
         self.swagger_types = {
             'security_context_type': 'str',
             'run_as_user': 'int',
             'run_as_group': 'int',
             'is_non_root_user_check_enabled': 'bool',
-            'is_root_file_system_readonly': 'bool'
+            'is_root_file_system_readonly': 'bool',
+            'capabilities': 'ContainerCapabilities'
         }
 
         self.attribute_map = {
@@ -56,7 +61,8 @@ class LinuxSecurityContext(SecurityContext):
             'run_as_user': 'runAsUser',
             'run_as_group': 'runAsGroup',
             'is_non_root_user_check_enabled': 'isNonRootUserCheckEnabled',
-            'is_root_file_system_readonly': 'isRootFileSystemReadonly'
+            'is_root_file_system_readonly': 'isRootFileSystemReadonly',
+            'capabilities': 'capabilities'
         }
 
         self._security_context_type = None
@@ -64,6 +70,7 @@ class LinuxSecurityContext(SecurityContext):
         self._run_as_group = None
         self._is_non_root_user_check_enabled = None
         self._is_root_file_system_readonly = None
+        self._capabilities = None
         self._security_context_type = 'LINUX'
 
     @property
@@ -161,6 +168,26 @@ class LinuxSecurityContext(SecurityContext):
         :type: bool
         """
         self._is_root_file_system_readonly = is_root_file_system_readonly
+
+    @property
+    def capabilities(self):
+        """
+        Gets the capabilities of this LinuxSecurityContext.
+
+        :return: The capabilities of this LinuxSecurityContext.
+        :rtype: oci.container_instances.models.ContainerCapabilities
+        """
+        return self._capabilities
+
+    @capabilities.setter
+    def capabilities(self, capabilities):
+        """
+        Sets the capabilities of this LinuxSecurityContext.
+
+        :param capabilities: The capabilities of this LinuxSecurityContext.
+        :type: oci.container_instances.models.ContainerCapabilities
+        """
+        self._capabilities = capabilities
 
     def __repr__(self):
         return formatted_flat_dict(self)