+++ src/oci/osp_gateway/invoice_service_client.py
@@ -100,7 +100,7 @@ class InvoiceServiceClient(object):
             'regional_client': True,
             'service_endpoint': kwargs.get('service_endpoint'),
             'base_path': '/20191001',
-            'service_endpoint_template': 'https://ospap.oracle.com',
+            'service_endpoint_template': 'https://osp-oci-integ.osp.{region}.oci.{secondLevelDomain}',
             'service_endpoint_template_per_realm': {  },  # noqa: E201 E202
             'skip_deserialization': kwargs.get('skip_deserialization', False),
             'circuit_breaker_strategy': kwargs.get('circuit_breaker_strategy', circuit_breaker.GLOBAL_CIRCUIT_BREAKER_STRATEGY),